ai-photo/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
face-recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
